requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
pydantic
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
sqlalchemy